+++ packages/sdk/src/adaptors/boardroom.ts
@@ -13,7 +13,7 @@ interface BoardroomResult {
 }
 
 export default async function getBoardroomData(address: string) {
-  const data = (await fetcher(
+  const { data } = (await fetcher(
     'GET',
     `https://api.boardroom.info/v1/voters/${address}/votes`
   )) as BoardroomResult;
@@ -21,9 +21,12 @@ export default async function getBoardroomData(address: string) {
   let totalVotes = 0;
   const daos = [];
   const votes = [];
-  for (let index = 0; index < data['data'].length; index++) {
-    const doc = data['data'][index];
-    if (doc?.proposalInfo?.currentState === 'executed') {
+  for (let index = 0; index < data.length; index++) {
+    const doc = data[index];
+    if (
+      doc?.proposalInfo?.currentState === 'executed' ||
+      doc?.proposalInfo?.currentState === 'closed'
+    ) {
       totalVotes += 1;
       if (daos.includes(doc.protocol) === false) {
         daos.push(doc.protocol);