+++ packages/sdk/src/adaptors/context.ts
@@ -1,6 +1,7 @@
 import { fetcher } from '../utils';
 
 interface ContextResult {
+  error?: boolean;
   followerCount: number;
   followingCount: number;
 }
@@ -11,8 +12,11 @@ export default async function getContextData(address: string) {
     'https://context.app/api/profile/' + address
   )) as ContextResult;
 
-  return {
-    followerCount: data.followerCount,
-    followingCount: data.followingCount,
-  };
+  if (data?.error === true) return false;
+  else if (Object.keys(data).includes('followerCount') === false) return false;
+  else
+    return {
+      followerCount: data.followerCount,
+      followingCount: data.followingCount,
+    };
 }